refactor: clean tests
