+++ tests/units/typing/url/test_any_url.py
@@ -1,7 +1,6 @@
-import json
-
 from pydantic.tools import parse_obj_as, schema_json_of
 
+from docarray.document.io.json import orjson_dumps
 from docarray.typing import AnyUrl
 
 
@@ -16,6 +15,6 @@ def test_json_schema():
     schema_json_of(AnyUrl)
 
 
-def test_dump_json(json_encoder):
+def test_dump_json():
     url = parse_obj_as(AnyUrl, 'http://jina.ai/img.png')
-    json.dumps(url, cls=json_encoder)
+    orjson_dumps(url)
